rrr: convert rpc/reconnect_policy to module (rrr.reconnect_policy)
reconnect_policy.cpp is the rrr.reconnect_policy module interface unit;
.hpp deleted. ReconnectPolicy POD with 3 presets (aggressive /
conservative / no_retry) + ReconnectCalculator (exponential backoff +
jitter via std::random_device).

Includers (2) updated. 21 modules total.
